+++ src/main/java/fitnesbot/models/WaterGoal.java
@@ -2,5 +2,5 @@ package fitnesbot.models;
 
 import fitnesbot.services.NutrientUnits;
 
-public record WaterGoal(int quantity, NutrientUnits units) {
+public record WaterGoal(double quantity, NutrientUnits units) {
 }
